requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers for BERT & Wav2Vec2
transformers>=4.30.0

# Video & Audio Processing
opencv-python>=4.8.0  # Extract frames from videos
av>=10.0.0  # PyAV - video/audio reading
librosa>=0.10.0  # Audio analysis
soundfile>=0.12.0  # Audio I/O

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0

# Progress bars & utilities
tqdm>=4.65.0
pyyaml>=6.0

# Experiment Tracking
tensorboard>=2.13.0

# Development
jupyter>=1.0.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0  # ASGI server to run FastAPI
python-multipart>=0.0.6  # For file uploads

# Database
sqlalchemy>=2.0.0  # ORM for database operations
psycopg2-binary>=2.9.0  # PostgreSQL driver
pgvector>=0.2.0  # Vector similarity search
asyncpg>=0.28.0  # Async PostgreSQL driver
python-dotenv>=1.0.0  # Environment variables
